+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Fix blurhash zero width resizing bug @reebalazs
 - Fix ignoring blurhash width if width (or size) was a string @reebalazs
+- Update fast-blurhash with recent blurhash fixes @reebalazs
 
 ### Internal